+++ src/macros.rs
@@ -131,10 +131,28 @@ macro_rules! dynamic_uniform{
 ///
 /// ## Naming convention
 ///
-/// When it comes to using to using your vertex array in a shader you must make sure that all your attribute variables *match* the field names in the struct you are calling calling this macro for.
+/// If not using the location option, when it comes to using to using your vertex array in a shader you must make sure that all your attribute variables *match* the field names in the struct you are calling calling this macro for.
 ///
-/// So, if you have a `vertex_position` atribute/input in your shader, a field named `vertex_position` must be present in the struct. Ohterwise the drawing functions will panic.
+/// So, if you have a `vertex_position` attribute/input in your shader, a field named `vertex_position` must be present in the struct. Otherwise the drawing functions will panic.
 ///
+/// ## Normalize option
+///
+/// You can specify a normalize option for attributes.
+/// ```
+/// # use glium::implement_vertex;
+/// # fn main() {
+/// implement_vertex!(Vertex, position normalize(false), tex_coords normalize(false));
+/// # }
+/// ```
+/// ## Location option
+///
+/// You can specify a location option for attributes.
+/// ```
+/// # use glium::implement_vertex;
+/// # fn main() {
+/// implement_vertex!(Vertex, position location(0), tex_coords location(1));
+/// # }
+/// ```
 #[macro_export]
 macro_rules! implement_vertex {
     ($struct_name:ident, $($field_name:ident),+) => (
@@ -150,6 +168,7 @@ macro_rules! implement_vertex {
                         (
                             Cow::Borrowed(stringify!($field_name)),
                             $crate::__glium_offset_of!($struct_name, $field_name),
+                            -1,
                             {
                                 // Obtain the type of the $field_name field of $struct_name and
                                 // call get_type on it.
@@ -182,6 +201,7 @@ macro_rules! implement_vertex {
                         (
                             Cow::Borrowed(stringify!($field_name)),
                             $crate::__glium_offset_of!($struct_name, $field_name),
+                            -1,
                             {
                                 // Obtain the type of the $field_name field of $struct_name and
                                 // call get_type on it.
@@ -196,6 +216,41 @@ macro_rules! implement_vertex {
                             {
                                 $should_normalize
                             }
+                        )
+                    ),+
+                ])
+            }
+        }
+    };
+
+    ($struct_name:ident, $($field_name:ident location($location:expr)),+) => {
+        impl $crate::vertex::Vertex for $struct_name {
+            #[inline]
+            fn build_bindings() -> $crate::vertex::VertexFormat {
+                use std::borrow::Cow;
+
+                // TODO: use a &'static [] if possible
+
+                Cow::Owned(vec![
+                    $(
+                        (
+                            Cow::Borrowed(stringify!($field_name)),
+                            $crate::__glium_offset_of!($struct_name, $field_name),
+                            {
+                                $location
+                            },
+                            {
+                                // Obtain the type of the $field_name field of $struct_name and
+                                // call get_type on it.
+                                fn attr_type_of_val<T: $crate::vertex::Attribute>(_: Option<&T>)
+                                    -> $crate::vertex::AttributeType
+                                {
+                                    <T as $crate::vertex::Attribute>::get_type()
+                                }
+                                let field_option = None::<&$struct_name>.map(|v| &v.$field_name);
+                                attr_type_of_val(field_option)
+                            },
+                            false
                         )
                     ),+
                 ])
@@ -238,7 +293,7 @@ macro_rules! implement_buffer_content {
                 type Owned = Box<$struct_name<$($gs)*>>;
 
                 #[inline]
-                fn read<F, E>(size: usize, f: F) -> ::std::result::Result<Box<$struct_name<$($gs)*>>, E>
+                unsafe fn read<F, E>(size: usize, f: F) -> ::std::result::Result<Box<$struct_name<$($gs)*>>, E>
                               where F: FnOnce(&mut $struct_name<$($gs)*>) -> ::std::result::Result<(), E>
                 {
                     use std::mem;
@@ -465,7 +520,7 @@ macro_rules! implement_uniform_block {
 ///         vertex: r#"
 ///             #version 300
 ///
-///             fn main() {
+///             void main() {
 ///                 gl_Position = vec4(0.0, 0.0, 0.0, 1.0);
 ///             }
 ///         "#,
@@ -473,7 +528,7 @@ macro_rules! implement_uniform_block {
 ///             #version 300
 ///
 ///             out vec4 color;
-///             fn main() {
+///             void main() {
 ///                 color = vec4(1.0, 1.0, 0.0, 1.0);
 ///             }
 ///         "#,
@@ -482,14 +537,14 @@ macro_rules! implement_uniform_block {
 ///         vertex: r#"
 ///             #version 110
 ///
-///             fn main() {
+///             void main() {
 ///                 gl_Position = vec4(0.0, 0.0, 0.0, 1.0);
 ///             }
 ///         "#,
 ///         fragment: r#"
 ///             #version 110
 ///
-///             fn main() {
+///             void main() {
 ///                 gl_FragColor = vec4(1.0, 1.0, 0.0, 1.0);
 ///             }
 ///         "#,
@@ -498,14 +553,14 @@ macro_rules! implement_uniform_block {
 ///         vertex: r#"
 ///             #version 110
 ///
-///             fn main() {
+///             void main() {
 ///                 gl_Position = vec4(0.0, 0.0, 0.0, 1.0);
 ///             }
 ///         "#,
 ///         fragment: r#"
 ///             #version 110
 ///
-///             fn main() {
+///             void main() {
 ///                 gl_FragColor = vec4(1.0, 1.0, 0.0, 1.0);
 ///             }
 ///         "#,