+++ src/vertex_array_object.rs
@@ -259,10 +259,22 @@ impl VertexArrayObject {
     {
         // checking the attributes types
         for &(_, ref bindings, _, _, _) in vertex_buffers {
-            for &(ref name, _, ty, _) in bindings.iter() {
-                let attribute = match program.get_attribute(Borrow::<str>::borrow(name)) {
-                    Some(a) => a,
-                    None => continue
+            for &(ref name, _, location, ty, _) in bindings.iter() {
+                let attribute = match location {
+                    -1 => {
+                        // No location specified in Vertex Format. Check name instead
+                        match program.get_attribute(Borrow::<str>::borrow(name)) {
+                            Some(a) => a,
+                            None => continue,
+                        }
+                    }
+                    _ => {
+                        match program.attributes().into_iter()
+                                .find(|(_, a)| a.location == location) {
+                            Some((_, a)) => a,
+                            None => continue,
+                        }
+                    }
                 };
 
                 if ty.get_num_components() != attribute.ty.get_num_components() ||
@@ -274,11 +286,23 @@ impl VertexArrayObject {
             }
         }
 
+        // checking for duplicate attribute locations
+        for &(_, ref bindings, _, _, _) in vertex_buffers {
+            for (i, bi) in bindings.iter().enumerate() {
+                for (o, bo) in bindings.iter().enumerate() {
+                    if i != o && bi.2 == bo.2 && bi.2 != -1 {
+                        panic!("The program attribute `{}` has the same binding location as program attribute `{}` (binding location {})",
+                               bi.0, bo.0, bi.2)
+                    }
+                }
+            }
+        }
+
         // checking for missing attributes
-        for (&ref name, _) in program.attributes() {
+        for (&ref name, attribute) in program.attributes() {
             let mut found = false;
             for &(_, ref bindings, _, _, _) in vertex_buffers {
-                if bindings.iter().any(|&(ref n, _, _, _)| n == name) {
+                if bindings.iter().any(|&(ref n, _, location, _, _)| (location != -1 && location == attribute.location) || n == name) {
                     found = true;
                     break;
                 }
@@ -520,12 +544,24 @@ unsafe fn bind_attribute(ctxt: &mut CommandContext<'_>, program: &Program,
     }
 
     // binding attributes
-    for &(ref name, offset, ty, normalize) in bindings.iter() {
+    for &(ref name, offset, location, ty, normalize) in bindings.iter() {
         let (data_type, elements_count, instances_count) = vertex_binding_type_to_gl(ty);
 
-        let attribute = match program.get_attribute(Borrow::<str>::borrow(name)) {
-            Some(a) => a,
-            None => continue
+        let attribute = match location {
+            -1 => {
+                // No location specified in Vertex Format. Check name instead
+                match program.get_attribute(Borrow::<str>::borrow(name)) {
+                    Some(a) => a,
+                    None => continue,
+                }
+            }
+            _ => {
+                match program.attributes().into_iter()
+                        .find(|(_, a)| a.location == location) {
+                    Some((_, a)) => a,
+                    None => continue,
+                }
+            }
         };
 
         if attribute.location != -1 {
@@ -552,6 +588,14 @@ unsafe fn bind_attribute(ctxt: &mut CommandContext<'_>, program: &Program,
                                                         elements_count as gl::types::GLint, data_type, 0,
                                                         stride as i32,
                                                         (buffer_offset + offset + (i * elements_count * 4) as usize) as *const _)
+                        }
+                    },
+                    gl::HALF_FLOAT => {
+                        for i in 0..instances_count {
+                            ctxt.gl.VertexAttribPointer((attribute.location + i) as u32,
+                                                        elements_count as gl::types::GLint, data_type, 0,
+                                                        stride as i32,
+                                                        (buffer_offset + offset + (i * elements_count * 2) as usize) as *const _)
                         }
                     },